backend/requirements.txt
fastapi
uvicorn
google-generativeai
slowapi
python-dotenv
